app/python/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
backoff==2.2.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
exceptiongroup==1.2.0
Flask==2.3.3
googleapis-common-protos==1.62.0
grpcio==1.60.0
h11==0.14.0
httptools==0.6.1
idna==3.6
importlib-metadata==6.11.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
opentelemetry-api==1.22.0
opentelemetry-distro==0.43b0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-aws-lambda==0.43b0
opentelemetry-instrumentation-dbapi==0.43b0
opentelemetry-instrumentation-flask==0.43b0
opentelemetry-instrumentation-jinja2==0.43b0
opentelemetry-instrumentation-logging==0.43b0
opentelemetry-instrumentation-sqlite3==0.43b0
opentelemetry-instrumentation-tortoiseorm==0.43b0
opentelemetry-instrumentation-urllib==0.43b0
opentelemetry-instrumentation-wsgi==0.43b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
packaging==23.2
protobuf==4.25.1
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.32.0.post1
typing_extensions==4.9.0
urllib3==2.1.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==2.3.8
wrapt==1.16.0
zipp==3.17.0
